third_party/blink/web_tests/http/tests/security/isolatedWorld/bypass-main-world-csp-for-xhr-expected.txt

CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
CONSOLE ERROR: Connecting to 'http://127.0.0.1:8000/security/isolatedWorld/resources/empty.html' violates the following Content Security Policy directive: "connect-src 'none'". The action has been blocked.
Tests that isolated worlds can have XHRs that the page's CSP wouldn't allow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

XHR from main world
PASS: The request was disallowed
XHR from isolated world with unchanged CSP
PASS: The request was disallowed
XHR from isolated world with looser CSP
PASS: The request succeeded
XHR from main world is not affected by the isolated world origin or CSP
PASS: The request was disallowed
PASS successfullyParsed is true

TEST COMPLETE
